feat(ChatList)/#3: ChatList 컴포넌트 분리, 채팅방 더미데이터 추가
